Further refactoring
- compiling the model is defined now in a separate script, compile_model.py
- the live_predict_translate and google_apis are moved to a separate folder
- the live_predict_translate now has a parameter to define if a user wants to use Google APIs to translate the recording or is just interested in the predictions made by neural network

+++ record_predict/live_predict_translate.py
@@ -6,14 +6,15 @@ import argparse
 import subprocess
 import soundfile as sf
 import sounddevice as sd
+from tensorflow.keras.models import load_model
 
 from google_apis import text_to_speech, transcribe_speech, translate_text
 
-lib_dir = os.path.join(os.getcwd(), "./tensorflow")
+lib_dir = os.path.join(os.getcwd(), "../tensorflow")
 sys.path.append(lib_dir)
 from SpectrogramGenerator import SpectrogramGenerator
 from predict import predict
-from load_compile_model import load_compile_model
+from compile_model import compile_model
 
 
 def count_down():
@@ -32,7 +33,7 @@ def play(audio_file_path):
     p.communicate()
 
 
-def live_predict_translate(model_dir, config):
+def live_predict_translate(model_dir, use_google_apis, config):
     fs = 44100 #Hz
     duration = 3 #seconds
 
@@ -50,36 +51,36 @@ def live_predict_translate(model_dir, config):
         print('\nIdentifying the language...')
         print('The proportions below are in the following order: Croatian, French, Spanish')
         result = predict(model_dir, filename, config)
+        
+        if use_google_apis:
+            try:
+                transcribed = transcribe_speech(filename, result)
+                translated = translate_text(transcribed)
+                output = 'output.mp3'
+                text_to_speech(translated, output)
+                print('Playing the translation!')
+                play(output)
+                os.remove(output)
+            except Exception as e:
+                print('\nThe following exception occured:')
+                print(e)
 
-        try:
-            transcribed = transcribe_speech(filename, result)
-            translated = translate_text(transcribed)
-            output = 'output.mp3'
-            text_to_speech(translated, output)
-            print('Playing the translation!')
-            play(output)
-            os.remove(output)
-        except Exception as e:
-            print('\nThe following exception occured:')
-            print(e)
-
-            print("--------------------------\n")
-            print("Note from the author: If it's not a Google API error,")
-            print("it may just be that the sounddevice wasn't able to access the microphone")
-            print("--------------------------\n")
-        os.remove(filename)
-        #os.remove(downsampled)
-
+                print("--------------------------\n")
+                print("Note from the author: If it's not a Google API error,")
+                print("it may just be that the sounddevice wasn't able to access the microphone")
+                print("--------------------------\n")
+            os.remove(filename)
+            
         print('\nWould you like to record and identify again? [y/n]: ')
         to_record = input()
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument('--model', dest='model_dir', required=True)
-    parser.add_argument('--config', dest='config', default="tensorflow/config.yaml")
+    parser.add_argument('--config', dest='config', default="../tensorflow/config.yaml")
+    parser.add_argument('--google_apis', dest='google_apis', default=0,
+                        type=int, help="Pass '1' if you'd like to use google APIs for translation, otherwise ignore")
     args = parser.parse_args()
-
-    model = load_compile_model(args.model_dir)
-    live_predict_translate(model, args.config)
-    
-    
+    model = load_model(args.model_dir)
+    model = compile_model(model)
+    live_predict_translate(model, args.google_apis, args.config)
